feat(weaves): mid-game weave learning via teacher NPC + PP (Phase E #28)
Adds the practice-points spend path for learning new weaves outside
chargen. Migration 0043 introduces weave_teachers (1:1 to mob_template,
mirrors trainers) carrying max_level_taught + affinity_filter PowerSet;
optional `weave_teacher:` YAML block on a mob seeds the row.

Earning: ComputeLevelUp.PracticeDelta = 1/level for every class,
deposited via RecordLevelUp into characters.practice_points (column
from migration 0009 was previously unwritten). Spend: `learn weave`
now branches on room context — teacher present drains practice_points
through the new RecordWeaveStudy (mirrors RecordWeavePick's tx shape
but writes practice_points instead of pending_weaves) and applies
the teacher's level cap + affinity filter intersected with the
channeler's own affinities; teacher absent keeps the existing
pending_weaves chargen-pool path verbatim. Audit on the mid-game
path: kind=weave_study power=<p> cost=<n>.

Catalog: chargen.Weave gained PracticeCost (validated [0,255]); all
eight level-0 weaves seed practice_cost: 1. score's Channeling block
gained a Practice line. PracticePoints widened to int32 to match the
SQLite INTEGER column. commitWeaveStudy enforces a teacher-filter
defense-in-depth gate so callers bypassing the menu still hit the
allowed list.

RecordWeaveStudy shares the same RMW TOCTOU as RecordWeavePick;
tracked in optimistic_lock_followups.md for the multi-session batch.

+++ internal/creature/creature.go
@@ -719,7 +719,7 @@ type Character struct {
 	XP             int64
 	Feats          []int32
 	Skills         map[int32]SkillRanks
-	PracticePoints int16
+	PracticePoints int32
 	ClassFeatures  []int32
 
 	HeightCm   int16
